requirements.txt
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
alembic==1.12.0
asyncpg==0.28.0
